requirements.txt
alembic==1.13.1
anyio==4.2.0
appdirs==1.4.4
backoff==2.2.1
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
commonmark==0.9.1
contourpy==1.2.0
cycler==0.12.1
dacite==1.6.0
dagshub==0.3.12
databricks-cli==0.18.0
dataclasses-json==0.6.3
docker==6.1.3
entrypoints==0.4
Flask==3.0.0
fonttools==4.47.2
fusepy==3.0.1
gitdb==4.0.11
GitPython==3.1.41
gql==3.5.0
graphql-core==3.2.3
gunicorn==21.2.0
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.6
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.3
joblib==1.3.2
kiwisolver==1.4.5
Mako==1.3.0
Markdown==3.5.2
MarkupSafe==2.1.3
marshmallow==3.20.2
matplotlib==3.8.2
mlflow==2.9.2
multidict==6.0.4
mypy-extensions==1.0.0
mysql-connector-python==8.3.0
numpy==1.26.3
oauthlib==3.2.2
packaging==23.2
pandas==2.1.4
pathvalidate==3.0.0
pillow==10.2.0
protobuf==4.25.2
pyarrow==14.0.2
Pygments==2.17.2
PyJWT==2.8.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
querystring-parser==1.2.4
requests==2.31.0
requests-toolbelt==1.0.0
rfc3986==1.5.0
rich==13.1.0
scikit-learn==1.3.2
scipy==1.11.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
SQLAlchemy==2.0.25
sqlparse==0.4.4
tabulate==0.9.0
tenacity==8.2.3
threadpoolctl==3.2.0
treelib==1.6.4
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
websocket-client==1.7.0
Werkzeug==3.0.1
xgboost==2.0.2
yarl==1.9.4
zipp==3.17.0
